Fix CI test failures: Add missing parameter and dtype compatibility
- Add explicit enable_openequivariance parameter to MACEStack.__init__()
- Fix TypeError: __init__() got unexpected keyword argument 'enable_openequivariance'
- Change test dtype from float64 to float32 for e3nn compatibility
- Resolve TorchScript tensordot dtype mismatch errors in CI tests
- Maintain backward compatibility with default enable_openequivariance=False"

+++ tests/test_openequivariance_numerical_equivalence.py
@@ -412,7 +412,9 @@ def pytest_configuration_consistency(enable_oe):
     irreps_out = o3.Irreps("1x0e")
 
     linear = OptimizedLinear(irreps_in, irreps_out)
-    x = torch.randn(4, irreps_in.dim, dtype=torch.float64)
+    x = torch.randn(
+        4, irreps_in.dim, dtype=torch.float32
+    )  # Use float32 for compatibility
 
     with torch.no_grad():
         y = linear(x)
